pyproject.toml
[project]
name = "mcp-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black==24.10.0",
    "chromadb==1.2.1",
    "colorama>=0.4.6",
    "docx2md==1.0.4",
    "exceptiongroup>=1.0.0rc8",
    "fastmcp==2.12.5",
    "httpx==0.28.1",
    "iniconfig>=2.3.0",
    "isort==5.13.2",
    "llama-index-core==0.14.5",
    "mcp==1.16.0",
    "openai==2.6.0",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "pluggy>=0.12,<2.0",
    "polars==1.35.2",
    "pre-commit==4.3.0",
    "psycopg2-binary==2.9.11",
    "pylint==3.3.1",
    "pytest==7.4.3",
    "pytest-asyncio>=0.23.8",
    "pytest-mock==3.14.0",
    "python-dotenv==1.1.1",
    "scikit-learn>=1.8.0",
    "tavily-python==0.7.13",
    "tomli>=1.0.0",
    "typing-extensions>=4.6.0",
]
